ishemalink_api/requirements.txt
Django==4.2.10
djangorestframework==3.14.0
drf-spectacular==0.27.1
psycopg2-binary==2.9.9
redis==5.0.1
python-decouple==3.8
django-filter==23.5
django-cors-headers==4.3.1
gunicorn==21.2.0
